更新数据模型，加入按装置信息，增加总报警率计算

+++ AlarmManage/Models/ViewRate.cs
@@ -45,6 +45,6 @@ namespace AlarmManage.Models
         /// <summary>
         /// 时间范围秒数
         /// </summary>
-        public int Range { get; set; }
+        public double Range { get; set; }
     }
 }
